third_party/WebKit/LayoutTests/platform/chromium-mac/fast/forms/file/input-file-value-expected.txt

This tests the value attribute in file input forms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL file.value should be C:\fakepath\foo.txt. Was C:\fakepath\.
PASS file.files.length is 1
PASS file.value = 'foo' threw exception Error: InvalidStateError: DOM Exception 11.
FAIL file.value should be C:\fakepath\foo.txt. Was C:\fakepath\.
PASS file.files.length is 1
PASS file.value is ""
PASS file.files.length is 0
PASS successfullyParsed is true

TEST COMPLETE
